NeverballPack: add OBJ as a model format
I didn't do this originally because it made models hard to
select. Doing this, however, allows for a neat trick: set "modelscale"
to "64" and "angles" to "0 0 90" on the misc_model entity to have the
model actually show up in Radiant correctly (provided that Radiant has
OBJ support, which NetRadiant does have).

git-svn-id: https://s.snth.net/svn/neverball/trunk@3631 78b8d119-cf0a-0410-b17c-f493084dd1d7